move col names
