+++ projects/_04_direct_payment_recipients/jobs/merge_dpr_data.py
@@ -1,7 +1,7 @@
 from pyspark.sql import DataFrame
 
 from utils import utils
-from utils.column_names.direct_payments_column_names import (
+from projects._04_direct_payment_recipients.tests.utils.direct_payments_column_names import (
     DirectPaymentColumnNames as DP,
 )
