+++ projects/_04_direct_payment_recipients/jobs/split_pa_filled_posts_into_icb_areas.py
@@ -6,7 +6,7 @@ from utils import utils, cleaning_utils
 from utils.column_names.cleaned_data_files.ons_cleaned import (
     OnsCleanedColumns as ONSClean,
 )
-from utils.column_names.direct_payments_column_names import (
+from projects._04_direct_payment_recipients.tests.utils.direct_payments_column_names import (
     DirectPaymentColumnNames as DPColNames,
 )
 from projects._04_direct_payment_recipients.direct_payments_configuration import (
