+++ projects/_04_direct_payment_recipients/tests/utils/_01_prepare_dpr_utils/test_dpr_prepare_during_year_data.py
@@ -11,7 +11,7 @@ from pyspark.sql.types import (
 
 from utils import utils
 import projects._04_direct_payment_recipients.utils._01_prepare_dpr_utils.prepare_during_year_data as job
-from utils.column_names.direct_payments_column_names import (
+from projects._04_direct_payment_recipients.tests.utils.direct_payments_column_names import (
     DirectPaymentColumnNames as DP,
 )
