+++ projects/_04_direct_payment_recipients/tests/utils/_03_estimate_direct_payment_utils/test_dpr_calculate_remaining_variables.py
@@ -11,7 +11,7 @@ from pyspark.sql.types import (
 
 from utils import utils
 import projects._04_direct_payment_recipients.utils._03_estimate_direct_payment_utils.calculate_remaining_variables as job
-from utils.column_names.direct_payments_column_names import (
+from projects._04_direct_payment_recipients.tests.utils.direct_payments_column_names import (
     DirectPaymentColumnNames as DP,
 )
