+++ projects/_04_direct_payment_recipients/utils/_03_estimate_direct_payment_utils/estimate_service_users_employing_staff.py
@@ -1,6 +1,6 @@
 from pyspark.sql import DataFrame, functions as F
 
-from utils.column_names.direct_payments_column_names import (
+from projects._04_direct_payment_recipients.tests.utils.direct_payments_column_names import (
     DirectPaymentColumnNames as DP,
 )
 from projects._04_direct_payment_recipients.utils._03_estimate_direct_payment_utils.models.extrapolation_ratio import (
