+++ projects/_04_direct_payment_recipients/utils/_03_estimate_direct_payment_utils/models/interpolation.py
@@ -3,7 +3,7 @@ import sys
 from pyspark.sql import DataFrame, Window, functions as F
 from pyspark.sql.types import ArrayType, LongType, FloatType
 
-from utils.column_names.direct_payments_column_names import (
+from projects._04_direct_payment_recipients.tests.utils.direct_payments_column_names import (
     DirectPaymentColumnNames as DP,
 )
